pyproject.toml
[project]
name = "invermin-system"
version = "0.1.0"
description = "Sistema Comercial para Invermin Paititi"
readme = "README.md"
requires-python = ">=3.11"
# Mueve aquí las dependencias principales para que uv las instale por defecto
dependencies = [
    "fastapi>=0.132.0",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "pyodbc>=5.3.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "weasyprint>=68.1",
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "pg8000>=1.31.5",
    "psycopg2-binary>=2.9.11",
    "pre-commit>=4.5.1",
]

[dependency-groups]
# Grupo específico para desarrollo y migraciones
dev = [
    "alembic>=1.18.4",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear (recomendado)
]
ignore = [
    "E501", # ruff format ya se encarga del ancho de línea
    "B008",   # Depends() en defaults — patrón estándar FastAPI
    "F403",   # star imports en __init__.py de modelos
    "F405",   # nombres de star imports
]

[tool.ruff.lint.per-file-ignores]
"backend/alembic/*" = ["F401"]
"backend/alembic/env.py" = ["E402"]   # imports después de os.environ — necesario en alembic
"backend/app/models/base.py" = ["N805"]   # declared_attr usa cls, no self
